fixing merge

+++ Assets/Scripts/DataPersistence/Data/GameData.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [System.Serializable]
 public class GameData
 {
-    public float coinsCollected;
+    public int coinsCollected;
 
     // the values defined in this constructor will be the default values
     // the game starts with when there's no data to load
